refactoring tests by adding a subtest dir
