+++ packages/service-library/tests/aiohttp/test_application_setup.py
@@ -7,8 +7,8 @@ from unittest.mock import Mock
 
 import pytest
 from aiohttp import web
-from servicelib.application_keys import APP_CONFIG_KEY
-from servicelib.application_setup import (
+from servicelib.aiohttp.application_keys import APP_CONFIG_KEY
+from servicelib.aiohttp.application_setup import (
     APP_SETUP_KEY,
     DependencyError,
     ModuleCategory,
