+++ packages/service-library/tests/aiohttp/test_sandbox.py
@@ -3,7 +3,7 @@
 # pylint:disable=redefined-outer-name
 
 import pytest
-from servicelib import openapi
+from servicelib.aiohttp import openapi
 
 
 @pytest.fixture
